+++ Martina/DataTransferObjects/CreateRoomRequest.cs
@@ -14,11 +14,11 @@ public class CreateRoomRequest
     /// 房间的单价
     /// </summary>
     [Required]
-    public float PricePerDay { get; set; }
+    public decimal PricePerDay { get; set; }
 
     /// <summary>
     /// 房间的基础环境温度
     /// </summary>
     [Required]
-    public float RoomBasicTemperature { get; set; }
+    public decimal RoomBasicTemperature { get; set; }
 }
